task_1/requirements.txt
blinker==1.6.3
click==8.1.7
colorama==0.4.6
dnspython==2.4.2
Flask==3.0.0
Flask-WTF==1.2.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
numpy==1.26.1
pandas==2.1.2
pymongo==4.5.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
six==1.16.0
tzdata==2023.3
Werkzeug==3.0.1
WTForms==3.1.0
xlrd==2.0.1
